+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 <!-- There is always Unreleased section on the top. Subsections (Add, Changed, Fix, Removed) should be Add as needed. -->
 ## Unreleased
 - Update dependencies
+- Allow to `addRoute` to the _metrics_ WebServer
 
 ## 7.1.0 - 2019-08-08
 - Allow `registerCustomMetric` directly